@@ -3,11 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^18.0.0",
+    "react": "^18.2.0",
     "react-calendar-heatmap": "../",
-    "react-dom": "^18.0.0",
-    "react-scripts": "^5.0.0",
-    "react-tooltip": "^3.6.1"
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "react-tooltip": "^5.26.3"
   },
   "scripts": {
     "predeploy": "npm run build",
